// dashboard/src/types/sales.ts
export interface SalesTransaction {
  id: string;
  brand_name: string;
  product_division: string;
  product_group: string;
  product_category: string;
  date: string;
  item: string;
  qty_item: number;
  localamount: number;
  store_code: string;
  month: string;
  year: string;
  sap_article: string;
  sap_description: string;
  brand_code: string;
  lob: string;
  sublob: string;
  kpi_group: string;
  store_name: string;
  concept: string;
  week_apple: string;
  quarter_apple: string;
  week_sf: string;
}